+++ src/show.c
@@ -2,5 +2,5 @@
 #include "show.h"
 
 void show_int(void *data) {
-	printf("%d ", (int)(data));
+	printf("%d ", (int)(unsigned long)data);
 }
